+++ lib/features/borrower/domain/repositories/borrower_repository.dart
@@ -1,5 +1,5 @@
 import 'package:dartz/dartz.dart';
-import 'package:photon/core/error/failure.dart';
+import 'package:photon/core/error/failures.dart';
 import 'package:photon/features/borrower/domain/entities/borrower.dart';
 
 abstract class BorrowerRepository {
