+++ lib/features/borrower/domain/usecases/get_borrower.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:photon/core/error/failure.dart';
+import 'package:photon/core/error/failures.dart';
 import 'package:photon/core/usecases/usecase.dart';
 import 'package:photon/features/borrower/domain/entities/borrower.dart';
 import 'package:photon/features/borrower/domain/repositories/borrower_repository.dart';
